AVLabels default
